Update role options

+++ app/Http/Controllers/Admin/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 class UsersController extends Controller
@@ -32,7 +33,10 @@ class UsersController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(User $user)
-    {
+    {   
+        if(Gate::denies('edit-users')){
+            return redirect()->route('admin.users.index');
+        }
         $roles = Role::all();
         
         return view('admin.users.edit')->with([
@@ -52,6 +56,10 @@ class UsersController extends Controller
     {
         $user->roles()->sync($request->roles);
 
+        $user->name = $request->name;
+        $user->email = $request->email;
+        $user->save();
+
         return redirect()->route('admin.users.index');
     }
 
@@ -63,6 +71,9 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('delete-users')){
+            return redirect()->route('admin.users.index');
+        }
         $user->roles()->detach();
         $user->delete();
